feat: move cli tools to their own packages
- update cli docs
- fix #1659 close / disconnect error for serialport-terminal
- rename serialport-term to serialport-terminal
- stop `test:require` because it's not necessary and breaks on the cli packages
