+++ docs/guide-cli.md
@@ -3,19 +3,35 @@ id: guide-cli
 title: Command Line Tools
 ---
 
-If you install `serialport` globally (e.g., `npm install -g serialport`), you'll receive two command line tools. These can also be run with `npx` if serialport is a dependency in your project.
+(These cli tools were formally part of the `serialport` package and have now been moved to their own packages for ease of use.)
+
+All cli tools can be run via [`npx`](https://www.npmjs.com/package/npx) or installed globally.
 
 For Example:
 ```bash
-$ npm install --save serialport
-$ npx serialport-list
+$ npx @serialport/list
+/dev/tty.Bluetooth-Incoming-Port
+/dev/tty.usbmodem1421    Arduino (www.arduino.cc)
+
+#or
+
+$ npm install -g @serialport/list
+$ serialport-list
 /dev/tty.Bluetooth-Incoming-Port
 /dev/tty.usbmodem1421    Arduino (www.arduino.cc)
 
 ```
 
-## `serialport-list`
-`serialport-list` will list all available serial ports in different formats.
+## SerialPort List
+
+```bash
+$ npx @serialport/list [options]
+# or
+$ npm install -g @serialport/list
+$ serialport-list [options]
+```
+
+The package `@serialport/list` will install the `serialport-list` cli tool which lists all available serial ports in different formats.
 
 ```bash
 $ serialport-list -h
@@ -43,35 +59,15 @@ $ serialport-list -f jsonline
 {"comName":"/dev/tty.usbmodem1421","manufacturer":"Arduino (www.arduino.cc)","serialNumber":"752303138333518011C1","locationId":"14200000","vendorId":"2341","productId":"0043"}
 ```
 
-## `serialport-term`
-`serialport-term` provides a basic terminal interface for communicating over a serial port. `ctrl+c` will exit.
-
+## SerialPort Repl
 ```bash
-$ serialport-term -h
-
-  Usage: serialport-term -p <port> [options]
-
-  A basic terminal interface for communicating over a serial port. Pressing ctrl+c exits.
-
-  Options:
-
-    -h, --help                     output usage information
-    -V, --version                  output the version number
-    -l --list                      List available ports then exit
-    -p, --port, --portname <port>  Path or name of serial port
-    -b, --baud <baudrate>          Baud rate default: 9600
-    --databits <databits>          Data bits default: 8
-    --parity <parity>              Parity default: none
-    --stopbits <bits>              Stop bits default: 1
-    --echo --localecho             Print characters as you type them
-
-$ serialport-term -l
-/dev/tty.Bluetooth-Incoming-Port
-/dev/tty.usbmodem1421    Arduino (www.arduino.cc)
+$ npx @serialport/repl <port>
+# or
+$ npm install -g @serialport/repl
+$ serialport-repl <port>
 ```
 
-## `serialport-repl`
-`serialport-repl` provides a nodejs repl for working with serialport. This is valuable when debugging.
+The package `@serialport/repl` will install the `serialport-repl` cli tool which provides a nodejs repl for working with serialport. This is valuable when debugging.
 
 You can make use of the `serialport-repl` command with;
 ```bash
@@ -102,4 +98,38 @@ null
 > port.open()
   serialport:main opening path: /dev/tty.usbmodem1421 +30s
   serialport:bindings open +1ms
+```
+
+## SerialPort Terminal
+```bash
+$ npx @serialport/terminal  -p <port> [options]
+# or
+$ npm install -g @serialport/terminal
+$ serialport-terminal  -p <port> [options]
+```
+
+The package `@serialport/terminal` will install the `serialport-terminal` cli tool which provides a basic terminal interface for communicating over a serial port. `ctrl+c` will exit.
+
+```bash
+$ serialport-terminal -h
+
+  Usage: serialport-terminal -p <port> [options]
+
+  A basic terminal interface for communicating over a serial port. Pressing ctrl+c exits.
+
+  Options:
+
+    -h, --help                     output usage information
+    -V, --version                  output the version number
+    -l --list                      List available ports then exit
+    -p, --port, --portname <port>  Path or name of serial port
+    -b, --baud <baudrate>          Baud rate default: 9600
+    --databits <databits>          Data bits default: 8
+    --parity <parity>              Parity default: none
+    --stopbits <bits>              Stop bits default: 1
+    --echo --localecho             Print characters as you type them
+
+$ serialport-terminal -l
+/dev/tty.Bluetooth-Incoming-Port
+/dev/tty.usbmodem1421    Arduino (www.arduino.cc)
 ```
