+++ packages/list/package.json
@@ -1,0 +1,23 @@
+{
+  "name": "@serialport/list",
+  "version": "1.0.0",
+  "main": "lib/list.js",
+  "bin": {
+    "serialport-list": "./lib/list.js"
+  },
+  "dependencies": {
+    "commander": "^2.13.0",
+    "@serialport/bindings": "^2.0.2"
+  },
+  "engines": {
+    "node": ">=6.0.0"
+  },
+  "publishConfig": {
+    "access": "public"
+  },
+  "license": "MIT",
+  "repository": {
+    "type": "git",
+    "url": "git://github.com/node-serialport/node-serialport.git"
+  }
+}
