+++ packages/repl/package.json
@@ -1,0 +1,23 @@
+{
+  "name": "@serialport/repl",
+  "version": "1.0.0",
+  "main": "lib/repl.js",
+  "bin": {
+    "serialport-repl": "./lib/repl.js"
+  },
+  "dependencies": {
+    "promirepl": "^1.0.1",
+    "serialport": "^7.0.2"
+  },
+  "engines": {
+    "node": ">=6.0.0"
+  },
+  "publishConfig": {
+    "access": "public"
+  },
+  "license": "MIT",
+  "repository": {
+    "type": "git",
+    "url": "git://github.com/node-serialport/node-serialport.git"
+  }
+}
