+++ packages/terminal/.npmignore
@@ -1,0 +1,2 @@
+*.test.js
+CHANGELOG.md
