+++ packages/terminal/package.json
@@ -1,0 +1,24 @@
+{
+  "name": "@serialport/terminal",
+  "version": "1.0.0",
+  "main": "lib/terminal.js",
+  "bin": {
+    "serialport-terminal": "./lib/terminal.js"
+  },
+  "dependencies": {
+    "commander": "^2.13.0",
+    "prompt-list": "^3.2.0",
+    "serialport": "^7.0.2"
+  },
+  "engines": {
+    "node": ">=6.0.0"
+  },
+  "publishConfig": {
+    "access": "public"
+  },
+  "license": "MIT",
+  "repository": {
+    "type": "git",
+    "url": "git://github.com/node-serialport/node-serialport.git"
+  }
+}
